add the ui
This is a pretty basic UI that still do not interact with the
rss engine. It has functionality to mark/unmark a item bold/italic
when is un/read, hability to display the html text from a rss item
description and the hability to add top level items to the tree view
where the channels will be displayed.