extract out channel/feed functions to a class
This marks a major shift in how we handle the feeds. Now instead
of a single json with all channel and feed data, we store a file
with all channel data, and the we create another json with the
newly implemented channel uuid to store its contents inside.

As of now, the class has no way to access any serialized feed data,
only the channel data. This will be reimpemented in later commits.