implement tests
This will be useful when i start the migration from json to
sqllite. Because of the tests, the ChannelList classs had to
be modified to accept a new folder to store its data so we
can mock it without touching on the user files and now we
check if the folder exists before trying to delete the files.